divide several .js
