+++ frontend/my-app/src/App.css
@@ -64,11 +64,8 @@
   background-color: aqua;
 }
 
-.result{
 
-}
-
-.horizontalLayout{
+.horizontaliLayout{
   height: 750px;
 }
