+++ frontend/my-app/src/Content.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import CodeInput from './CodeInput';
+import HorizontalLayout from './HorizontalLayout';
+
+function Content(){
+    return <div className="content">
+      <CodeInput></CodeInput>
+      <HorizontalLayout></HorizontalLayout>
+    </div>
+  }
+
+
+export default Content;
+
