+++ frontend/my-app/src/Header.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+import Login from './Login';
+
+function Header(){
+    return <div className="header">
+      <Login></Login>
+      <div className="siteName">사이트 이름</div>
+    </div>
+  }
+
+export default Header;
